Fix to avoid exceeding db username field
svn path=/archetypes/trunk/; revision=4351

+++ openvpms-archetypes/src/test/java/org/openvpms/archetype/test/TestHelper.java
@@ -219,7 +219,8 @@ public class TestHelper extends Assert {
      * @return a new user
      */
     public static User createUser() {
-        return createUser("zuser" + System.nanoTime(), true);
+        // use an int to avoid exceeding the length of the db field
+        return createUser("zuser" + (int) System.nanoTime(), true);
     }
 
     /**
